Move useExtension to hooks dir
